clippy: drop literal tab in admin1 doc-comment + redundant i64 cast
The admin1 module schema example had a real `\t` character embedded
in the doc comment, which clippy::tabs_in_doc_comments rejected.
Replaced with `<TAB>` placeholder so the schema reads the same but
the file is tab-free. Also removed an `as i64` cast in the chirps
monthly tslot snap — `days_from_civil` already returns `i64`.

+++ crates/emem-api-rest/src/lib.rs
@@ -21933,7 +21933,7 @@ async fn materialize_chirps_monthly_precip(
     // Snap to first-of-month instant in days.
     let first_of_month_days = days_from_civil(year, month, 1);
     let tslot = emem_core::tslot::Tslot::from_unix(
-        (first_of_month_days as i64) * 86_400,
+        first_of_month_days * 86_400,
         emem_core::tslot::Tempo::Medium,
     )
     .0;
